requirements.txt
crewai
langchain-google-genai
python-dotenv
requests
pillow
pandas
tabulate
